% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataToy.R
\name{selectiveToy}
\alias{selectiveToy}
\alias{selectiveToy.rct}
\alias{selectiveToy.rwe}
\title{Toy Continuous Outcome Dataset}
\format{
selectiveToy provides two datasets. The selectiveToy.rct
2986 participant records; selectiveToy.rwe 1014 participant records. Each
data.frame provides the following:
\itemize{
\item Y: A continuous outcome.
\item X1: A continuous covariate.
\item X2: A continuous covariate.
\item A: A binary treatment variable
}
}
\usage{
data("selectiveToy")
}
\description{
These datasets are provided only to facilitate examples. They are not based
on or representative of any real-world applications.
}
\keyword{datasets}
