% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srEC.R
\name{srEC}
\alias{srEC}
\title{Selective and robust external-control borrowing strategy for evaluating treatment effect}
\usage{
srEC(
  data_rt,
  data_ec = NULL,
  rt.ctrl = caret::trainControl(method = "cv", number = 10),
  hc.ctrl = caret::trainControl(method = "cv", number = 10),
  method = "gbm",
  ...
)
}
\arguments{
\item{data_rt}{A list contains X, Y and A for RT. The propensity scores P(A=1|X) can also be
contained as \code{prob_A} (or \code{NULL}).}

\item{data_ec}{A list contains X and Y for EC with treatment A=0.}

\item{rt.ctrl}{An object of \code{\link[caret:trainControl]{caret::trainControl()}} for \code{\link[caret:train]{caret::train()}}, which controls the
computational nuances for fitting the outcome model using the RT dataset.}

\item{hc.ctrl}{An object of \code{\link[caret:trainControl]{caret::trainControl()}}, which controls the
computational nuances for fitting the outcome model using the EC dataset.}

\item{...}{Other options used to fit the predictive models. Passed on to \code{\link[caret:train]{caret::train()}}.}
}
\value{
A list with components:
\itemize{
\item est: estimated average treatment effect by AIPW, ACW and the selective integrative estimator.
\item sd: estimated standard errors for the aforementioned estimators.
\item subset.idx: a subset of indices of the external controls which have been selected for the
final integrative estimation.
}
}
\description{
\code{srEC()} is a penalized dynamic integrative framework to augment a randomized trial (RT)
with a external control (EC) dataset, in which the subject-level compatibility of the EC
is assessed by a well-crafted penalty (e.g., adaptive lasso penalty). The parameter of interest
is the average treatment effect.
}
