% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srEC_Surv.R
\name{srEC_Surv}
\alias{srEC_Surv}
\title{Selective and robust external-control borrowing strategy for evaluating treatment effect}
\usage{
srEC_Surv(
  data_rt,
  data_ec,
  k_grid,
  X.pred.R = c("X1", "X2", "X3"),
  X.pred.A = c("X1", "X2", "X3"),
  X.pred.c = c("X1", "X2", "X3"),
  X.pred.f = c("A", "X1", "X2", "X3")
)
}
\arguments{
\item{data_rt}{A list contains X, Y, D and A for RT. The propensity scores P(A=1|X) can also be
contained as \code{prob_A} (or \code{NULL}).}

\item{data_ec}{A list contains X, Y, and D for EC with treatment A=0.}

\item{k_grid}{A vector of time points to evaluate the survival curves.}

\item{X.pred.R}{A list of covariate name to model the propensity for R}

\item{X.pred.A}{A list of covariate names to model the propensity for A}

\item{X.pred.c}{A list of covariate names to model the survival curve of the event}

\item{X.pred.f}{A list of covariate names to model the survival curve of the censoring}
}
\value{
A list with components:
\itemize{
\item aipcw.RCT: RCT-only estimate
\item aipcw: RCT and full EC estimate
\item aipcw_lasso: RCT with selective EC estimate
\item subset.idx: a subset of indices of the external controls which have been selected for the
final integrative estimation.
}
}
\description{
\code{srEC_Surv()} is a penalized dynamic integrative framework for survival outcomes that augments
a randomized trial (RT) with a external control (EC) dataset. The subject-level compatibility of the EC
is assessed by a well-crafted penalty (e.g., adaptive lasso penalty). The parameter of interest
is the average treatment effect in terms of the restricted mean survival time (RMST).
}
